+++ src/evo/synaptogenesis_rules.py
@@ -15,12 +15,177 @@
 # ==============================================================================
 
 import logging
+import traceback
 from evo import voxels
 from inf import runtime_data
 from random import randrange
+from math import prod
 
 
 logger = logging.getLogger(__name__)
+
+
+def neighbor_finder(cortical_area_src, cortical_area_dst, src_neuron_id):
+    """
+    Finds a list of candidate Neurons from another Cortical area to build Synapse with for a given Neuron
+    """
+
+    # Candidate_voxel_list includes a list of destination neuron and associated postSynapticCurrent pairs
+    candidate_voxel_list = list()
+    raw_candidate_list = set()
+
+    # rule_manager = SynaptogenesisRuleManager(src_neuron_id=src_neuron_id, src_cortical_area=cortical_area_src,
+    #                                          dst_cortical_area=cortical_area_dst)
+    # candidate_list = rule_manager.growth_rule_selector()
+
+    src_voxel = runtime_data.brain[cortical_area_src][src_neuron_id]['soma_location']
+
+    morphologies = runtime_data.genome["blueprint"][cortical_area_src]['cortical_mapping_dst'][cortical_area_dst]
+
+    for morphology_ in morphologies:
+        neuron_morphology = morphology_['morphology_id']
+        morphology_scalar = morphology_['morphology_scalar']
+        psc_multiplier = morphology_['postSynapticCurrent_multiplier']
+        psc_base = runtime_data.genome["blueprint"][cortical_area_src]['postsynaptic_current']
+        post_synaptic_current = psc_multiplier * psc_base
+
+        try:
+            for key in runtime_data.genome["neuron_morphologies"][neuron_morphology]:
+                # print("Morphology:", key)
+                if key == "vectors":
+                    for vector in runtime_data.genome["neuron_morphologies"][neuron_morphology]["vectors"]:
+                        candidate_list = match_vectors(src_voxel=src_voxel, cortical_area_dst=cortical_area_dst,
+                                                       vector=vector, morphology_scalar=morphology_scalar)
+                        if candidate_list:
+                            for candidate in candidate_list:
+                                raw_candidate_list.add((candidate[0], candidate[1], candidate[2]))
+                            # candidate_voxel_list.append([matching_vectors, post_synaptic_current])
+                        candidate_list = None
+
+                elif key == "patterns":
+                    for pattern in runtime_data.genome["neuron_morphologies"][neuron_morphology]["patterns"]:
+
+                        candidate_list = match_patterns(src_voxel=src_voxel, cortical_area_dst=cortical_area_dst,
+                                                        pattern=pattern, morphology_scalar=morphology_scalar)
+                        if candidate_list:
+                            for candidate in candidate_list:
+                                raw_candidate_list.add((candidate[0], candidate[1], candidate[2]))
+                                # candidate_voxel_list.append([item, post_synaptic_current])
+                        candidate_list = None
+                elif key == "functions":
+                    if neuron_morphology == "expander_x":
+                        candidate_list = syn_expander_x(cortical_area_src, cortical_area_dst, src_neuron_id)
+                        for candidate in candidate_list:
+                            raw_candidate_list.add((candidate[0], candidate[1], candidate[2]))
+                            # candidate_voxel_list.append([candidate, post_synaptic_current])
+                    elif neuron_morphology == "reducer_x":
+                        candidate_list = syn_reducer_x(cortical_area_src, cortical_area_dst, src_neuron_id)
+                        for candidate in candidate_list:
+                            raw_candidate_list.add((candidate[0], candidate[1], candidate[2]))
+                            # candidate_voxel_list.append([candidate, post_synaptic_current])
+                    elif neuron_morphology == "randomizer":
+                        candidate = syn_randomizer(dst_cortical_area=cortical_area_dst)
+                        raw_candidate_list.add((candidate[0], candidate[1], candidate[2]))
+                        # candidate_voxel_list.append([candidate, post_synaptic_current])
+                    elif neuron_morphology == "lateral_pairs_x":
+                        candidate = syn_lateral_pairs_x(neuron_id=src_neuron_id, cortical_area=cortical_area_src)
+                        raw_candidate_list.add((candidate[0], candidate[1], candidate[2]))
+                        # candidate_voxel_list.append([candidate, post_synaptic_current])
+                    elif neuron_morphology == "block_connection":
+                        candidate = syn_block_connection(cortical_area_src, cortical_area_dst, src_neuron_id, s=10)
+                        raw_candidate_list.add((candidate[0], candidate[1], candidate[2]))
+                        # candidate_voxel_list.append([candidate, post_synaptic_current])
+                    elif neuron_morphology == "projector":
+                        candidate_list = syn_projector(cortical_area_src, cortical_area_dst, src_neuron_id)
+                        for candidate in candidate_list:
+                            raw_candidate_list.add((candidate[0], candidate[1], candidate[2]))
+                            # candidate_voxel_list.append([candidate, post_synaptic_current])
+                    candidate_list = None
+                elif key == "placeholder":
+                    pass
+
+                else:
+                    print("Warning! Morphology %s did not have any valid definition." % neuron_morphology)
+
+        except Exception as e:
+            print("Error during synaptogenesis of %s and %s" % (cortical_area_src, cortical_area_dst))
+            print(traceback.format_exc())
+
+        for candidate in raw_candidate_list:
+            candidate_voxel_list.append([list(candidate), post_synaptic_current])
+
+    if candidate_voxel_list:
+        candidate_neuron_list = \
+            voxels.voxel_list_to_neuron_list(cortical_area=cortical_area_dst,
+                                             voxel_list=candidate_voxel_list)
+        return candidate_neuron_list
+
+
+def match_vectors(src_voxel, cortical_area_dst, vector, morphology_scalar):
+    scaled_vector = [prod(x) for x in zip(vector, morphology_scalar)]
+    candidate_vector = [sum(x) for x in zip(src_voxel, scaled_vector)]
+    for item in candidate_vector:
+        if item < 0:
+            return None
+    within_limits = voxels.block_size_checker(cortical_area=cortical_area_dst,
+                                              block=voxels.block_reference_builder(candidate_vector))
+    if within_limits:
+        return [candidate_vector]
+
+
+def match_patterns(src_voxel, cortical_area_dst, pattern, morphology_scalar):
+    """
+    Matches source voxels to destination voxels
+
+    Expected pattern format:    [source pattern, destination pattern] e.g. [["*", "?", 3], [2, "*", "?"]]
+
+    """
+    voxel_list = list()
+    dst_block_boundaries = runtime_data.genome["blueprint"][cortical_area_dst]["block_boundaries"]
+
+    if len(pattern) != 2:
+        print("Error! Pattern was not defined correctly.. "
+              "should be similar to e.g. [[\"*\", \"?\", 3], [2, \"*\", \"?\"]]\n Current is as:", pattern)
+
+    src_pattern_x, src_pattern_y, src_pattern_z = pattern[0]
+    dst_pattern_x, dst_pattern_y, dst_pattern_z = pattern[1]
+
+    src_x, src_y, src_z = src_voxel
+
+    for dst_x in range(dst_block_boundaries[0]):
+        for dst_y in range(dst_block_boundaries[1]):
+            for dst_z in range(dst_block_boundaries[2]):
+
+                matching_condition_x = \
+                    ((dst_pattern_x == "*" and
+                      (src_pattern_x == "*" or src_pattern_x == "?" or src_pattern_x == src_x)) or
+                     (dst_pattern_x == "?" and (src_pattern_x == "*" or src_pattern_x == "?" or src_x == dst_x)) or
+                     (dst_pattern_x == dst_x and (src_pattern_x == "*" or
+                                                  (src_pattern_x == "?" and src_x == dst_x) or
+                                                  (src_pattern_x == src_x))))
+
+                matching_condition_y = \
+                    ((dst_pattern_y == "*" and
+                      (src_pattern_y == "*" or src_pattern_y == "?" or src_pattern_y == src_y) or
+                      (dst_pattern_y == "?" and (src_pattern_y == "*" or src_pattern_y == "?" or src_y == dst_y)) or
+                      (dst_pattern_y == dst_y and (src_pattern_y == "*" or (src_pattern_y == "?" and src_y == dst_y) or
+                                                  (src_pattern_y == src_y)))))
+
+                matching_condition_z = \
+                    ((dst_pattern_z == "*" and
+                      (src_pattern_z == "*" or src_pattern_z == "?" or src_pattern_z == src_z) or
+                      (dst_pattern_z == "?" and (src_pattern_z == "*" or src_pattern_z == "?" or src_z == dst_z)) or
+                      (dst_pattern_z == dst_z and (src_pattern_z == "*" or (src_pattern_z == "?" and src_z == dst_z) or
+                                                  (src_pattern_z == src_z)))))
+
+                if matching_condition_x and matching_condition_y and matching_condition_z:
+                    voxel_list.append([dst_x, dst_y, dst_z])
+
+    # print("Matched voxel list based on pattern:", src_voxel, cortical_area_dst, voxel_list)
+
+    # todo: account for morphology scalar
+
+    return voxel_list
 
 
 def syn_expander_x(src_cortical_area, dst_cortical_area, src_neuron_id, dst_y_index=0, dst_z_index=0):
